requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.1.0
bcrypt==3.2.0
bson==0.5.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
colorlog==4.2.1
configparser==5.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==3.1
DateTime==4.3
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-filters==0.2.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
enum34==1.1.10
filelock==3.0.12
filetype==1.0.7
har2case==0.3.1
HttpRunner==1.5.8
idna==2.7
importlib-metadata==1.7.0
iniconfig==1.1.1
item==0.0.3
itypes==1.2.0
Jinja2==2.11.2
jsonpath==0.82
jsonschema==3.2.0
loguru==0.5.1
lxml==4.6.1
MarkupSafe==1.1.1
more-itertools==8.5.0
namedlist==1.8
numpy==1.19.1
packaging==20.4
paramiko==2.7.2
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.11.0
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
Pypubsub==4.0.3
pyrsistent==0.16.0
pytest==4.5.0
python-dateutil==2.8.1
pytils==0.3
pytz==2020.1
pywin32==228
PyYAML==5.3.1
requests==2.20.0
requests-toolbelt==0.9.1
robotframework==3.2.2
robotframework-mongodb-library==3.2
robotframework-pythonlibcore==2.1.0
robotframework-requests==0.7.1
robotframework-ride==1.7.4.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.5.0
robotframework-sshlibrary==3.4.0
scp==0.13.2
selection==0.0.14
selenium==3.141.0
sentry-sdk==0.13.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tox==3.19.0
uritemplate==3.0.1
urllib3==1.24.3
user-agent==0.1.9
virtualenv==20.0.31
wcwidth==0.2.5
weblib==0.1.30
win32-setctime==1.0.3
wxPython==4.0.7.post2
xlrd==1.2.0
yagmail==0.11.224
zipp==3.1.0
zope.interface==5.2.0
